pkg: Moving topology module under pkg/
The topology module does not contain resource policy related
files so moving it back to pkg/
