+++ cmd/topology-aware/policy/flags.go
@@ -16,7 +16,7 @@ package topologyaware
 
 import (
 	config "github.com/containers/nri-plugins/pkg/config"
-	"github.com/containers/nri-plugins/pkg/resmgr/topology"
+	"github.com/containers/nri-plugins/pkg/topology"
 )
 
 // Options captures our configurable policy parameters.
