+++ cmd/topology-aware/policy/node.go
@@ -17,8 +17,8 @@ package topologyaware
 import (
 	"fmt"
 
-	"github.com/containers/nri-plugins/pkg/resmgr/topology"
 	system "github.com/containers/nri-plugins/pkg/sysfs"
+	"github.com/containers/nri-plugins/pkg/topology"
 	idset "github.com/intel/goresctrl/pkg/utils"
 	"k8s.io/kubernetes/pkg/kubelet/cm/cpuset"
