+++ pkg/resmgr/cache/container.go
@@ -25,7 +25,7 @@ import (
 	"github.com/containers/nri-plugins/pkg/cgroups"
 	"github.com/containers/nri-plugins/pkg/kubernetes"
 	"github.com/containers/nri-plugins/pkg/resmgr/apis"
-	"github.com/containers/nri-plugins/pkg/resmgr/topology"
+	"github.com/containers/nri-plugins/pkg/topology"
 
 	"github.com/containerd/nri/pkg/api"
 	nri "github.com/containerd/nri/pkg/api"
